Run Marklogic in its own container as part of script/server
When Django is being run natively via `script/server`, we want Marklogic to
run in its own Docker container. Add a `docker-marklogic.yml` file to build
and run this container, and run `docker-compose up` as part of script/server

The Marklogic instance is run in the background so that we can view the Django
logs in the terminal. Marklogic logs are piped to `marklogic.log`

Co-Authored-By: Lorna Harwood <lorna@dxw.com>