Change the way we import data into Marklogic for testing
Before, we were importing some test data into Marklogic with an import script.
Now we have a way of restoring our test database from a backup (hosted on S3 -
TODO, add AWS credentials)

Amend the boostrap script to set up the Application server in Marklogic (if it
does not already exist) ready to serve data to the front end.