Allow use of Marklogic to be switched off in development
When developing the front end of the application we do not need to be using
Marklogic - we can instead read XML files from the filesystem to keep the
applications more lightweight for the frontend developers.

Introduce a USE_MARKLOGIC environment variable, which when False, will cause
the API client to read XML files from the filesystem, instead of retrieving
them from Marklogic.