Add Index page to application
Add an index page to the application, with basic pagination.

In Marklogic-using mode, this will retrieve judgments 10 at a time and present
them in a simple list with pagination.

In mocked Marklogic mode, this will present 2 pages of static search results,
retrieved from the fixtures directory.

Because of difficulty in mocking a multipart/mixed content type, we have
created a more simple view  in non-Marklogic mode, which does not include the
Judgment names. In a future interation we could create a mocked multipart/mixed
response and parse it in the same way as in the Marklogic mode.