Import home template from flask app
I have commented out all url tags within the template files as they are
not defined yet and is out of scope for this work. This also applies
to the form within the basic_search_form template.
I have also commented out the original `for` loop syntax within
browse_by_court template to keep as reference.