Import Results template from Flask app
I have commented out the form within results_applied_filters template as
this requires further changes to get it to work and is therefore out of
scope for this commit.

The index page url path has been amdended to only match with a digit,
as previously this was matching with anything that was within the url.
This meant that the results url was matching with the index url path by
mistake.