Update Judgment name
A simple proof-of-concept form to update a Judgment's name in its metadata.

This commit adds an admin app (`judgment-admin`) with an edit & update method
for Judgments. A judgment can be edited by visiting `/judgments_admin/{uri}`.
The edit form only exposes the metadata name  (Xpath: `//akn:FRBRname/@value`)
for editing.

When the form is submitted, the judgment XML is updated, and the XML document
is sent to Marklogic via a PUT method in the API client.