Client-side implementation of cookie consent
This brings in the HTML, CSS and JavaScript from the main TNA
cookie consent repository into a Django app. Where possible I've
isolated the code into its own directory with a README referencing
the main repository.

A comment has also been added to the JavaScript to show where the
TNA domain should be inserted once the application is on a TNA domain.