Use `api_client.get_judgment_name` to get a Judgment's metadata name
Previously, we were having to get the entire Judgment XML just so we could
parse it for the Judgment's name (to be used in the breadcrumbs).

The reason
we needed tot get the XML and not use the result of `api_client.eval_xslt` is
because the result of `api_client.eval_xslt` is HTML and the location of the
Judgment name is not obvious, whereas the result of `api_client.get_judgment_xml`
is XML and the metadata name is easily findable.

However, by using an XQuery and the `eval` endpoint, we can simply get the name
of the Judgment directly from Marklogic.