Do not show non-web images on Judgments
Some of the Judgmetns/decisions contain images which are not web formats, e.g.
.emf or .wmf

These formats are not viewable in a browser and cause an error when the
judgment is downloaded as a PDF.

Amend the XSL transform to only show web formats (png, gif, jpg) and ignore
any others.

Ideally, we will need Clerks to only submit images in judgments which are
web-compatible, or we will need to convert non-web images somewhow as part
of the ingestion service.

Note that the XSL transform will need to be re-uploaded to Marklogic as per the
README.