Introduce static browse by court (see below)
The team decided yesterday that as an interim measure the browse by
court functionality should be implemented on the front end. I have
therefore introduced the dictionary used in the Flask application
and provided this to the view.

Ultimately, this will need to be made dynamic by members of the back
end development team.