Show the correct (editable) Neutral Citation in the search results
An editor noted that they had edited the Neutral Citation for a judgment, which
had updated the metadata correctly, but the old citation was showing in the
search results.

It transpires we were picking out the incorrect element for the citation -
we were using the `akn:neutralCitation` element, which is not editable by
the editors. We should instead be picking out the citation from the judgment
metadata, `uk:cite`.