Add sorting to search results, by relevance or date descending
This is a temporary-ish fix until we get front-end design input.

Add Sort by date/Sort by relevance links to the search results.

The links re-write the request to alter the `order` query parameter. As such,
we have some code in the view to remove the `order` param and re-add it with
the relevant value.

We also have a tweak in the front end to not show the Order parameter in the
removable filter UI pattern.