Improve the query string passed to the paginator & "order by" template
Previously, we were always removing the `order` query string in the context, so
that when this query string was used by the "order by" fragment, it did not
already contain
an `order` parameter. However, this had the side effect of removing the `order`
parameter from the paginator, so only the first page of results was ordered
correctly (subsequent pages had default ordering).

Now, we have added a template filter to remove a named param from the query
string, and can use this filter to remove the `order` param in the template. This
does mean that we need to pass the `query_params` in the context for the
`results` view.