Handle Ctrl-C in 'fab run' and close Docker container
When Ctrl-C is pressed to stop the execution of the server, it would
instead stop the execution of the run() function. We now catch that explicitly.

In addition, we now stop the django Docker container we opened,
so that the user doesn't have to manually do something about the
still-running server which prevents new servers from running on the
old port.