Show a 404 if someone tries to view an unpublished judgment
If someone tries to hack a URL to get a Judgment that exists, but is unpublished
they receive a 500 Error. It would be "nicer" if they received a 404
instead.

This is the hacky quick first-pass solution. The AttributeError is thrown from
the MultipartDecoder, because it is trying to split something which is
"Unsplittable"