Adjust filter toggle approach to prevent FOUC
This commit introduces inline JavaScript to apply a `.js-enabled` class
to the body element and namespaced CSS to prevent the filters from
being shown where JavaScript is available. Updates are also made to the
`manage_filters` plugin to ensure that it presents the correct state (both
in terms of the button text and `collapsed` class) when the plugin first
runs and then when users have interacted with it.