Restrict FOUC changes to results page
When testing this morning I found that the CSS rule that is applied
to prevent a FOUC on the results page is also impacting the visibility
of the filters on the structured search page. This commit simply
increases the specificity to ensure the styles are not applied to the
structured search page.