Change where we pick a judgement's date from in the model
Back when we started developing this application, we weren't really aware of the
differences between `FRBRWork`, `FRBRManifestation` and `FRBRExpression`. When
picking the date to show when a judgment was handed down, we decided to choose
any `FRBRdate` element with an attribute of `judgment`.

As we have received more & more judgments, some of them have been displayed
in the public ui without a date. This is because not all the judgments have
an `FRBRdate/@name=judgment` in their XML.

Jim jurisdatum suggested that we look instead for ANY date in the `FRBRWork`
block, and use that as the judgment date. It might be a judgment date, a hearing
date etc. but it is the "correct" date for the work.

Co-authored-by: jurisdatum <jim@jurisdatum.com>