Move patching of AtomFeed to where search is imported
It feels like a constant issue in Python testing that @patch needs
to refer to the location that the patched thing is imported and
not where it is imported from

(so `from cats import kittens` in animals.py needs an @patch that
says `animals.kittens` not `cats.kittens`)

This particular case worked fine with patching the place it was
imported from. Mostly. Except when another test
invoked `self.client.get` before it. Which for some unknown reason
broke the patching and meant it was trying to send data to an
actual MarkLogic instance.

This makes no sense. If it makes sense to you, please let me know.