Use `npx` to run SASS
`npx` ensures that we're using the version of SASS installed via npm, which is controlled through our package.json, and not any other one that happens to be on the path of the machine in question. This fixes a specific problem in my dev environment (`asdf`, my ruby version manager used on a different project, seems to be jealously guarding the `sass` command and throwing an error), but also should contribute to greater portability in general!