Remove robots directives from static pages
This commit selectively removes the noindex,nofollow 'robots' meta tag from all the static pages, as requested in Trello ticket #922.

I've taken a hidden-by-default approach - the robots tag appears in the `base.html` template, and we use a jinja2 block to allow it to be overridden in the templates that extend it. In the static page templates we therefore override this with an empty block to allow them to be indexed.

I've included a test of the homepage (as an example of a page where we do not want the tag to be included) and the results page (as an example of a page where we want the tag to be included) in order to validate the effectiveness of this approach, however it seemed overkill to include a test for every single template. If you disagree however, i'd be happy to add them!