Fix 'back to top' UI bug.
As reported on trello ticket 391 - on short judgements (eg https://caselaw.nationalarchives.gov.uk/ewhc/comm/2011/68#js-phase-banner), the 'back to top' link appears above the site header.

This ticket fixes this by only displaying the link when the length of the judgement itself exceeds the size of the viewport.

(In addition, i've changed the .editorconfig file to respect what appears to be the norm in the existing code of useing 4 spaces for tabs in js files, without this my editor tends to mangle every file I touch with whitespace changes - hope this is ok!)

While this is a very small change I'd appreciate review, particularly on the following two points:

1) Is this an appropriate place to be using JQuery? I'm basing this decision on the recommendation in [TNA's developer gude](https://github.com/nationalarchives/front-end-development-guide/blob/master/development-guide.md) as things like viewport dimensions tend in my experience to be a cross-browser compatibility can of worms. Very happy to revise in vanilla JS if that's more appropriate here.

2) I'm not 100% happy with adding a second ".hidden" class to the link and managing the state of both a 'shown' and 'hidden' class independently - however, this is necessary to ensure that the 'back to top' link is not shown at all on short judgements (it's rather unsightly and redundant even at the bottom of the page) when JS is enabled, while also ensuring that it's shown and is functional when js is not enabled or not loaded.

Anyway, sorry for the ridiculous commit-message to lines-changed ratio, I'll be briefer about these things when I've got settled in :-)

Many thanks!

Tim