Fix django-admin errors
When  DJANGO_SETTINGS_MODULE is set in the environment for `fab sh`, the `django-admin` command fails as the given settings module is not on the python import path. resetting this so that it is empty fixes the problem (as seen also in `compose/production/django/start`, line 11). This commit ensures that this env var is reset on starting `fab-sh` so that django-admin commands can be run from within it without error!