Redesign download links to give XML less prominence
Many judges and clerks were confused by the 'download as XML' button, unsure what it was for, and doubly confused by the results of clicking it. Given it's probably a fairly niche use case, this probably suggests it had undue prominence in the previous design.

This is a proposal for a new design that moves the download links to the bottom of the page in a new 'download options' section, following Matt's design for the judgment summary page as much as posisble, and providing some contextual information for users to understand what the given download options are, and who they might be useful for. Given that downloading the judgement as a PDF is a very common use case, I've maintained the button for this at the top of the page, along with a subtle signposting link for other download options, which points to an anchor tag in the download options section in the footer.