First draft of Content Security Policy headers
We use the `django-csp` library to create CSP headers for us.

We permit a couple of Google URLs to provide script, style and fonts
to us: these can be found on the front page (and probably every page)

The inscrutible sha-256 hash `IWjjek`... refers to the inline script
in `templates/base.py` and appears to be the only inline script in
the public-ui.

There's two bigger problems:

* HTML versions of the judgments have a great many inline styles. These
are taken directly from the word document. We don't have a good way
of not using inline styles here right now.

* data: is used in the CSS, but isn't secure. It's fine, because the
CSS isn't dynamically generated, but I can't see an obvious way to
allow the CSS to use data:

We've set it to report only for now; it should report CSP issues
to rollbar as warnings. I recommend we run this in production for
a while, and if there's no errors, turn on the actual protection a
little later.