New design for advanced search filters
This reorders the fields to make sure that the most used filters are at the top, tweaks the design to make better use of space and make it a little easier to navigate, and removes a pair of fields which don't make sense in the context of the filter dropdown ('neutral citation' and 'specific keywords' - the former isn't a 'filter' as such, and the latter does the same thing as the search query). These are kept on the 'structured search' page becasue of specific expectations / needs of legal professionals.

I've also tweaked the structures of the templates to hopefully make things easier to understand / change.