Multiple court selections
This Converts the court select to a set of checkboxes for multiple selection. It's also involved brining in the court names and metadata across the site from the list exposed in ds-caselaw-utils version>=0.2.0, and various adjustments to query handling to make sure the list is properly passed through on removing filters and on paginating.

The styling of the checkboxes is recreated by hand from the GDS style guide - I was unable to use the prepacked GDS components because they depend fairly heavily on the base GDS styles in a way that causes various problems with the styles more generally!