Don't highlight stop words in search results
https://trello.com/c/Ry47hPz5

This change removes highlighting from individual stop words when they appear
*on their own* in search results. It won't remove highlighting from stop words
when they appear as part of a quoted search term, e.g. "evidence of wrongdoing"

This has to be in XSLT1.0 as that's what's used in the front end. Otherwise
the XPath would be neater because we could use `lower-case()` in XSLT2.0.