Require Judgment URIs to be /year/id terminated
Invalid paths like /2022/123/generated.pdf can be parsed as a
judgment uri with court 2022, year 123 and id generated.pdf,
leading to 500 errors. We tighten up the paths a little so they
have to have a 4-digit 'year' in the penultimate place and end
with a numeric 'identifier', so that we don't get the wrong information
in the wrong place.

It might be worth validating the judgment uri more closely (does
it start with a known court or court/subcourt combo before the numbers)