BUGFIX: in-page navigation links continue showing over header.
So, this turned out to be a horrible non-deterministic thing, so it took a while both to replicate and to track down. We use IntersectionObservers to test whether either the page header or footer is visible in the viewport, and hide the in page navigation links when that is the case. However, by using two seperate observers, we opened the door to a concurrency bug whereby if the 'footer' observer was called after the 'header' one, the links would be set to 'show', over the top of the header. Thankfully, IntersectionObservers can apply to multiple elements, so i've refactored to use a single observer on both elements which fixes the problem.