Removing search filters does not return to the top of the result list
This was probably an earlier oversight on my part- if you expand your search by removing a filter, it doesn't make sense to retain the page number you're currently on, as you're viewing an entirely different result set. This changes the behaviour to consistenly send you back to the first page when the scope of the search query is changed.