Do not pass stop words to unquoted search queries
If a search query is unquoted and contains stop words (the, and, of), do not pass
those stop words to the search query.

If the query is unquoted but only contains the word (the, of, and) then do not
remove the word, just search on it (why would you do that though!)

If the query is quoted (e.g. "weight of evidence") then do not remove the stop
words.