Use stop words stored in an external file
In a previous commit we added the stop words `the of and` to the application,
and prevented them from being passed to the search query. We now want to use a
larger set of stop words.

The words in `judgments/fixtures/stop_words.py` have been taken from
https://github.com/nltk/nltk

We have decided not to import NLTK because the TNA team would like to modify
the lift of stop words to suit their needs. By adding it to this repo, the
TNA team can edit `stop_words.py` themselves when they wish to add or remove
stop words.

The list of stop words is transformed into two regular expressions, which are
then used to filter the search query.