Fix empty results bug when browsing by court
When browsing by court from the homepage, and changing the number of results per page, no courts were returned. This was because the form to change the number of results per page was including Null parameters as hidden fields in error - have added a check to make sure these are not added to the query.