Add ability to cache court dates in the django postgres database
This stems from a discussion with Nick about ways of improving the current situation where we need to manually update the court metadata in the utils library every time a new judgment is published for a particular court in 2023.

Nick propose a cron job that populates the database with the start and end years for each court's current ingested range  on a daily or weekly basis - this handles the persistence layer in the app and the logic to choose the correct date range when displaying a court.