Add ability to cache court dates in the django postgres database
This stems from a discussion with Nick about ways of improving the current situation where we need to manually update the court metadata in the utils library every time a new judgment is published for a particular court in 2023.

Nick propose a cron job that populates the database with the start and end years for each court's current ingested range  on a daily or weekly basis - this PR is intended as a conversation starter on how that might work on the python side.

We add some thin decorators around the existing Court and CourtGroup classes from the utils library which check for the existence of the dates in the db, falling back to those defined in the yaml file if they don't exist.

This is a spike, so will need (a) some test coverage, and (b) some thinking about optimisation before it's ready for release. In particular, it currently does a single database query for each court in the metadata file, whenever the court metadata is needed, which is probably not ideal.

Comments, feedback, criticisms, counterproposals all welcome!