Fix regression in searching from homepage
PR #535 introduced a regression - in order to change the placeholder text based on the CSS breakpoint, we introduced two query input fields with different placeholders, which are set to display: none; selectively at different breakpoints.

The problem was, both had the same ID, so the second in the markup always overwrites the query parameter, meaning that the query term wasn't passed through in the request on desktop!

To fix this, I've renamed the mobile field to `query_m`, and made a small change in the results view to take the query from here if the desktop field is empty.