Fix format of date ranges supplied to date inputs
This fixes [#630 - Calendar year missing on firefox browers](https://trello.com/c/hG5D5ivY/630-calendar-year-missing-on-firefox-browsers) - the issue was that the `min` and `max` dates specified used the wrong format ("DD-MM-YYYY" instead of "YYYY-MM-DD"). It appears that most browsers were clever enough to parse the other format, but FF had troubles. Have amended to the YYYY-MM-DD format which is specified in the standard.