Fail tests fast if environment misconfigured
This particular test was failing for me locally as the assets host wasn't
set in my .env file, but the message was slightly cryptic so it took some time
(and some of @dragon-dxw's wisdom) to work out why.

This change makes the test less brittle (it always uses the configured assets host rather than being hardcoded), but also fails fast and prints a meaningful message if no assets host is configured at all, rather than spuriously passing as it checks for the presence of an empty string.