Search styles revamp.
This fixes a few problems that crept into the structured search page, as well as replacing the long placeholder text for a label, in consistent style, throughout all our search functionality. This gives consistency with the advanced search box, clearer signposting, and also gets rid of a nasty hack we were using to selectively include placeholder text based on screen size, which was causing no end of bugs (briefly - we needed two query inputs with different names, which needed to be dealt with on the server). I've also refactored the search field into a re-usable component, and deleted a load of duplicate styles!