Suppress WeasyPrint log spam
Weasyprint tells us a lot of information about the internal workings
of the PDF / CSS problems. We broadly don't care, and when people
are crawling the back-catalog of PDFs they trigger many lines of
logs per PDF hit, and we're hitting our Papertrail limit.

Adds SHOW_WEASYPRINT_LOGS environment variable, which is optional
  and may be set to True to allow the logs to print