Show 'order by' field even when no query string is present
Previously, we only showed the 'order by' dropdown when a query is present, as we only have a 'relevance' score when the query has performed a text match. However, this leads to two UX downsides when the query does not include a text string:

- No indication is given of how the results are ordered
- No option is available to switch between 'newest first' and 'oldest first' ordering

This corrects this - the 'relevance' option is only shown when a query string is provided, but the dropdown itself (with 'newest' and 'oldest' options) is always displayed.