Make npm ci fail on EBADENGINE
By default npm does not exit with an error code when the node version is not
the one that a package requires it merely warns that the engine is unsupported
and carries on without installing the package.

This means that Docker builds can succeed and we end up shipping broken images.

```
```

Adding `--engine-strict=true` to `npm ci` makes it fail if it encounters
EBADENGINE

```
```

Which then casues the docker build to fail and we wont ship bad images.