Add a model to handle PDF logic
This adds a new `DocumentPDF` class to handle all the things
we know about a PDF, including the URI and the size etc. I’ve
deliberately left off the presentation of the filesize, as
this will be handled in a seperate piece of work. All we’re
returning here is the raw data.

I’ve also rearranged the models so they now sit in seperate files