Pages for individual courts and tribunals
These list the judgments for that court with pagination, and also show a crest and summary blurb (the former stored in /static/images/court_crests, named with the court param with slashes substituted for underscores, the latter in /static/markdown/court_descriptions, named similarly). Courts without a crest or a description will pull in a default image (Currently the MoJ crest) and a default blurb (currently lorem ipsum) - these should be replaced with something suitable before release. The other option would be to omit crest / description where it is not provided.