Fix accidentally quadratic vs removal
We remove things like 'X - vs - Y' and 'X v Y' from search strings.
Github flagged up the previous code had a bug where if a large string was inserted,
the regex would take more and more time. (If one megabyte took
a second, ten might take a hundred or a million seconds.)

By rearranging the preprocess query so that we normalise the strings first,
we only ever need to look for a single space, but we also need to make sure
that if there isn't punctuation around the 'vs', we don't fix anything that
isn't broken (`evil` to -> `e v il` would be a mistake)

https://github.com/nationalarchives/ds-caselaw-public-ui/security/code-scanning/4