Include tna-frontend component library.
This was a little more involved than I thought - it required both an upgrade to the version of sass we're using, and some changes to the specificity of various rules (as tna-frontend seems to override the link colour in a manner that takes precedence over the CSS cascade in ways that are mysteriour to me).

This should however allow us to start including tna-frontend components in our pages though, and I think we should work towards further integration with tna-frontend, the next step would be to use the grid and the theming capability for layout, typography, and colours as much as possible.