Further fixes to visual breakage introduced by integtrating tna-frontend
A bit of a can of worms, this. Cristina noticed that footer links were the incorrect colour, in fixing that I noticed that the link-visited colour wasn't being applied without adding the global theme classname to the html element, which in turn reset some margins which broke various aspects of the layout subtley. I've addressed all of these as far as i can tell, and tested comprehensively (including checking a variety of judgments against production). we'll need all eyes on staging before we're ready for this to go live, but i think (fingers crossed) that this addresses the bulk of it. If we find any further issues i think we should reconsider using tna-frontend, at least for now (i'm going to raise an issue with them in order to ask them to scope their styles a little more closely).