Get more court data into the facet context
Rather than assume a very generic, meta-programmed approach to multiple
facets in the future, I am encoding a more specific approach to the only
facet we have so far, the court facet.

Including data about the court will allow us to get its human-friendly
display name, as well as the code used to formulate the query.

Signed-off-by: Cristina <579522+CristinaRO@users.noreply.github.com>