Include govuk-frontend JS in our bundle
This also requires us to switch to loading the bundle as a module - I have done this and tested that all our existing js still works!

In addition, I've added a CSS declaration to set the govuk font which gets rid of some spurious errors in the console as the default can't be loaded.