Fix review / submit steps.
The review step can't happen in the 'done' callback as the session storage of the form values is wiped on rendering, making it impossible to go back!

Instead we make a 'dummy step' with a custom template, and extend the WizardView to pass all the existing form data through to the template on each step.