CAF: licence holder defaults to main contact
A request from Chris Owens: when a user choses the "same contact" option
and doesnt provide a licence holder email or name, we duplicate the provided
main contact details in the licence_holder fields explicitly.

I've also taken the opportunity to refactor the data delivery part of
this somewhat in a way that should hopefully be clearer - the "xml" to
be forwarded to the licencing team is now explicitly set out in a
template which allows us to cut down on hackery in the python code.