Setup Marklogic for end-to-end tests in CI
This moves the end to end tests to a seperate action, and sets up
marklogic (as in the development environment), populating it with
some sample judgments from the live site, in order to run real
end-to-end tests.

It also includes a minimum-viable end-to-end test which exercises
marklogic (test_home_page.py) - marklogic is called to populate the
"recent judgments" on the homepage, so the homepage would 500 if it
was not available, and this test would fail.