perf(FCL-266): change behaviour of getting linked documents to allow response caching
When getting a linked document, previously it was always relying on a DocumentNotFoundError to reflect the absence of a linked document, but exceptions are not cached. Add a new parameter which allows the method to return None (a cachable response) in these cases where the absence of a document can be considered a somewhat harder failure, such as when looking up linked documents.