fix(FCL-278): allow homepage in robots.txt
At the moment, we disallow all URLs starting with `/` with a few exceptions for static pages. This includes disallowing the homepage. However, most search engines have decided that despite the disallow directive the page is important enough to list in the index anyway – but because of the disallow directive they can't scrape it to get anything like description snippets. This means although search engines reveal the existence of the site, they can't tell users anything useful about it. Change this behaviour to explicitly allow scraping of _only_ the root page.