Fix XML link
There are a number of places where we use formatted_document_uri to get the name of a resource -- images, PDFs and XML spring to mind.
But there's now a distinction between these resources. Those at `assets` will share the name of the MarkLogic resource, but
those at `/data.extension` will share the name of the loaded page.

Changing it only here seems correct.