Tidy up court and tribunal landing pages
These are no longer being used in the original way, and there is no path to get to them from anywhere else in the site (although they are listed in the sitemap). Retool these pages so they use the new court description content, and can be used as court-specific landing pages in future.